+++ axelor-supplier-portal/src/main/java/com/axelor/apps/supplierportal/service/ProductSupplierService.java
@@ -1,0 +1,9 @@
+package com.axelor.apps.supplierportal.service;
+
+import com.axelor.apps.base.db.Product;
+import com.axelor.apps.supplierportal.db.ProductSupplier;
+import com.axelor.exception.AxelorException;
+
+public interface ProductSupplierService {
+  public Product addOnCatalog(ProductSupplier productSupplier) throws AxelorException;
+}