+++ axelor-supplier-portal/src/main/java/com/axelor/apps/supplierportal/service/app/AppSupplierPortalServiceImpl.java
@@ -1,0 +1,20 @@
+package com.axelor.apps.supplierportal.service.app;
+
+import com.axelor.apps.base.db.AppSupplierPortal;
+import com.axelor.apps.base.db.repo.AppSupplierPortalRepository;
+import com.google.inject.Inject;
+
+public class AppSupplierPortalServiceImpl implements AppSupplierPortalService {
+
+  private AppSupplierPortalRepository appSupplierPortalRepo;
+
+  @Inject
+  public AppSupplierPortalServiceImpl(AppSupplierPortalRepository appSupplierPortalRepo) {
+    this.appSupplierPortalRepo = appSupplierPortalRepo;
+  }
+
+  @Override
+  public AppSupplierPortal getAppSupplierPortal() {
+    return appSupplierPortalRepo.all().fetchOne();
+  }
+}